requirements.txt
discord.py>=2.3.2
streamlit>=1.28.0
feedparser>=6.0.10
python-dotenv>=1.0.0
aiohttp>=3.9.0